+++ src/test/live-coding-playground/src/main/java/removefinal/RemoveFinalLocal.java
@@ -1,0 +1,8 @@
+package removefinal;
+
+public class RemoveFinalLocal {
+    public void met() {
+        final int x = 2;
+        System.out.println(x);
+    }
+}