+++ src/test/live-coding-playground/src/main/java/removefinal/RemoveFinalParam.java
@@ -1,0 +1,7 @@
+package removefinal;
+
+public class RemoveFinalParam {
+    public void met(final int x) {
+        System.out.println(x);
+    }
+}